rename entity folder
